Added retrieve and update crud service implementations

+++ src/main/java/lv/venta/service/IProductCRUDService.java
@@ -8,11 +8,11 @@ import lv.venta.model.ProductType;
 public interface IProductCRUDService {
 	void createProduct(String name, float price, int description, ProductType producttype) throws Exception;
 	
-	Product retrieveProductById(int id) throws Exception;
+	Product retrieveProductById(long id) throws Exception;
 	
 	ArrayList<Product> retrieveAllProduct() throws Exception;
 	
-	void updateProductById(String name, float price, int description, ProductType producttype, int id) throws Exception;
+	void updateProductById(String name, float price, int description, ProductType producttype, long id) throws Exception;
 	
-	void deleteProductById(int id) throws Exception;
+	void deleteProductById(long id) throws Exception;
 }
